Replaced time with time type

+++ src/core/types/time.py
@@ -1,1 +1,1 @@
-Time = int
+Time = float
